Issue #3078925 by robertragas, Kingdutch: Convert country code to Full name for better geocoding results
This commit removes a faulty fallback that undoes the conversion of the
country code that was added in a previous commit. This fixes the issue
described in #3078925 and makes the patch there unneccessary.

The fallback did nothing else than the function call already did and did
this when the function call was succesful instead of when it failed.