Remove geocoder dependency
The geocoder and geolocation module have a lot of overlap and using them
interchangeable can cause issues. The geolocation module has all the
features we need and is required for the Social Geolocation Maps module
to work, as well as helping us implement easier Search support in the
future. Therefor the Geocoder module is removed as a dependency so that
we reduce the total number of dependencies on sites using this module.