Introduce Social Geolocation Search module
The Social Geolocation Search module is introduced to hold the
configuration and alter hooks for supporting search in the Social
Geolocation module. Since the `search_api_location` module is required
for this to work it's better to move this into a 'connector' module
than to force users of the geolocation module to also have search
enabled.

This commit also adds the alter hook that makes geolocation search work
with SOLR. It's easier to keep an alter hook for each back-end than it
is to find a unified solution that works with both.

Some configuration changes have been made in how the geolocation module
interacts with the address field which causes the actual geocoding to
break.