Fix the configurability of the used geocoder
A configuration screen already existed to configure which geocoder
plugin should be used. However, this wasn't actually wired up to affect
any behaviour. The configuration is now used when loading the geocoding
and the keys of the configuration values have been chosen so that they
load the correct plugins.

Configuration options are now only added to the form if the plugin for
that option is enabled. This should help developers identify issues
where the correct module was not yet enabled.