Use AddressFormatter in address to string conversion
The logic in `_social_geolocation_address_to_string` is changed so use
the `DefaultFormatter` of the CommerceGuys/Addressing library. This
formatter takes the locale of the country of the address into account to
create a properly formatted address string.

This fixes an issue caused by Nominatim and Google interpreting
addresses from left to right by default expecting the country to be last
for Western countries.

Two examples that can be used in testing:

'Netherlands, Enschede' will geocode to an address in the United States
for a city called Netherlands. 'Enschede, Netherlands' will correctly
geocode to the city of Enschede in The Netherlands.

Both '4 Magnolia Ave., Santa Ana, CA, United States' and 'United States,
CA, Santa Ana, 4 Magnolia Ave.' will geocode correctly to the
corresponding address with the Google Geocoding API. However, using
Nominatim. only the former will result in a correctly geocoded address
and the country first format returns no results.

Using the formatter from this library also ensures that postal codes are
formatted correctly as the position of a postal code in an address
varies widely between countries and can lead to incorrect geocoding.