Fix Canada inconsistency with Postalcodes for Nominatim
This re-introduces a fix that was previously done in Issue #3086891 to
fix an inaccuracy with Nominatim that can lead to problems for Canadian
Postal codes. The issue is explained here: https://www.openstreetmap.org/user/lonvia/diary/43143

This fix was done in the Address Preprocessor which was removed because
Geocoder is no longer used.

Fixes #3086891
Fixes #3086892