Move query alter hook to views logic
The solr_query_alter hook is triggered too late in the search processing
which causes search_api_location options not to be applied. To resolve
this we hook into the views query altering which runs before the query
is handed to the API. By checking whether the query will be handled by
the Search API SOLR backend we achieve the same behavior.