Remove empty update hooks
The update hooks are no longer used. Drupal doesn't actually need update
hooks to exist in order and will automatically pick up the next highest
number. If an update hook no longer needs to run then it's safe to
remove.

The next update hook does still need to be 8004 because it will still
need to execute on platforms that may have already ran 8003.