Separate form handling for search and people overview
The people overview can work without search back-end so the logic for
altering the forms is split among the modules. The common logic for
adding things to the forms is extracted to a shared function and cleaned
up slightly.

The container type for the location filter is made configurable because
it should be a fieldset on the people overview and a details panel on
the search page. This fixes a visual bug that currently exists on the
people overview.