Remove geocoder dump configuration for address fields
Although the geocoder module can not be uninstalled because it may be
depended on by others, the configuration for the geocoding of Open
Social address field is removed. This geocoding is now handled by the
social_geolocation module using the geolocation module.

The post_update hook that adds the configuration is also removed as it
is no longer needed. The post_update hook that removes the configuration
should have no effect if the configuration isn't there because it's safe
to call `unset` on non-existant values.