Add new dependencies to composer
The Social Geolocation module will fail without the address module
present because we use classes from it. Additionally if the address
module ever removes the CommerceGuys/Addressing library then our module
will also break so we depend on it directly.

The geocoder module is not yet removed from the composer.json because
that may cause it to be removed from platforms without being uninstalled
in Drupal.