PoC implementation for fallback proximity calculation
This commit adds code that attempts to calculate the required radius to
search for all events within a chosen location. It does so by
calculating the cross-section of the bounding box returned from the
Nominatim geocoder and using half of it as the search radius.

This has not been tested to work with the Google Geocoding API yet.
Additionally the Nominatim API considers the Dutch Antillen to be a part
of The Netherland which means the bounding box for The Netherlands
covers 2/3rd of the world.

WARNING: THIS CODE IS JUST AN EXPERIMENT, IT CONTAINS FATAL FLAWS AND
SHOULD NOT BE USED IN ANY PROJECTS!