Fix Canada postal_code workaround
The `$address->get` returns a `TypedData` reference. This is actually a
class that has a reference to the value. When the value of the
`$address` field is changed then so does the value of the `TypedData`
object.

Additionally when trying to return the value from `''` to its original
value an error occurs. `TypedData` objects can not be converted to
strings directly but `->getValue()` should be called.

The fix is to retrieve a string value using `getValue()` from the typed
data. This does not change and can be properly passed to `set` on the
`$address` field when we're done with geocoding.